Update README.md
Bootstrap Studio v7.0.2

A powerful desktop app for creating responsive websites using the Bootstrap framework. Bootstrap Studio is a powerful desktop app for designing and prototyping websites. Bootstrap Studio is a desktop application that helps you create beautiful websites. It comes with a large number of built-in components, which you can drag and drop to assemble responsive web pages. It is built on top of the hugely popular Bootstrap framework, and exports clean and semantic HTML. Thousands of developers and designers use it every day. We are sure you'll love it too!

[Download Bootstrap Studio v7.0.2](https://developer.team/web-development/35120-bootstrap-studio-v702.html)